Support loading plugins with config.
Python-Markdown supports a number of extensions, and users of `blag` may
wish to utilize those beyond the defaults. This set of changes supports
loading those extensions by adding a line to the configuration file,
while adding some checks to make sure that an error is raised if a
non-existent extension is named.